backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# Background Tasks
celery==5.3.4
redis==5.0.1

# AI & Image Processing
google-generativeai==0.3.2
pillow==10.1.0
requests==2.31.0
# opencv-python-headless==4.8.1.78  # Temporarily disabled for easier deployment
numpy==1.24.3

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1

# OAuth Dependencies
python-jose[cryptography]==3.3.0
authlib==1.3.0
httpx==0.25.2
